+++ src/main/java/br/com/ifsp/tsi/bugtrackerbackend/service/TicketService.java
@@ -1,18 +1,29 @@
 package br.com.ifsp.tsi.bugtrackerbackend.service;
 
 import br.com.ifsp.tsi.bugtrackerbackend.dto.TicketDto;
+import br.com.ifsp.tsi.bugtrackerbackend.exception.TicketNotFoundException;
 import br.com.ifsp.tsi.bugtrackerbackend.model.entity.Ticket;
+import br.com.ifsp.tsi.bugtrackerbackend.model.entity.User;
 import br.com.ifsp.tsi.bugtrackerbackend.model.enums.TicketStatus;
 import br.com.ifsp.tsi.bugtrackerbackend.repository.TicketRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class TicketService {
+    private final UserService userService;
+    private final TicketCategoryService ticketCategoryService;
     private final TicketRepository ticketRepository;
 
-    public TicketService(TicketRepository ticketRepository) {
+    public TicketService(
+            UserService userService,
+            TicketCategoryService ticketCategoryService,
+            TicketRepository ticketRepository
+    ) {
+        this.userService = userService;
+        this.ticketCategoryService = ticketCategoryService;
         this.ticketRepository = ticketRepository;
     }
 
@@ -27,7 +38,12 @@ public class TicketService {
     }
 
     public Ticket createTicket(TicketDto request) {
-        var ticket = new Ticket(request);
+        var userDto = userService.getUserSignedIn();
+        var user = new User(userDto);
+
+        var category = ticketCategoryService.getCategoryById(request.categoryId());
+
+        var ticket = new Ticket(request, user, category);
 
         return ticketRepository.save(ticket);
     }
@@ -40,11 +56,12 @@ public class TicketService {
         return ticketRepository.save(ticket);
     }
 
-    public Ticket deleteTicket(Long ticketId) {
+    public void deleteTicket(Long ticketId) {
         var ticket = getTicketById(ticketId);
 
-        ticketRepository.delete(ticket);
+        if (ticket.getTicketId() != ticketId)
+            throw new TicketNotFoundException("Ticket not found", HttpStatus.NOT_FOUND);
 
-        return ticket;
+        ticketRepository.delete(ticket);
     }
 }
